requirements.txt


# # FastAPI Backend
# fastapi
# uvicorn
# python-multipart
# requests
# python-dotenv
# # LlamaIndex
# llama-index
# llama-index-llms-azure-openai
# llama-index-embeddings-azure-openai

# # Streamlit Frontend
# streamlit

fastapi
uvicorn
python-multipart
requests
python-dotenv
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
streamlit
python-magic